refactor: types 폴더 생성 및 기존 type 파일 이동
