+++ src/entities/post/types/index.ts
